AdventOfCode2020/AdventOfCode2020.Tests/Day14Tests.cs
using Xunit;

namespace AdventOfCode2020.Tests
{
	public class Day14Tests
	{
		private const string TestString1 = @"
mask = XXXXXXXXXXXXXXXXXXXXXXXXXXXXX1XXXX0X
mem[8] = 11
mem[7] = 101
mem[8] = 0";

		private const string TestString2 = @"
mask = 000000000000000000000000000000X1001X
mem[42] = 100
mask = 00000000000000000000000000000000X0XX
mem[26] = 1";

		private IDay Day { get; } = new Day14();

		[Theory]
		[InlineData(TestString1, 165)]
		public void Test1(string inputString, ulong expected)
		{
			var result = this.Day.Perform1(inputString);
			Assert.Equal(expected, result);
		}

		[Theory]
		[InlineData(TestString2, 208)]
		public void Test2(string inputString, ulong expected)
		{
			var result = this.Day.Perform2(inputString);
			Assert.Equal(expected, result);
		}
	}
}
